+++ changelog.md
@@ -3,6 +3,8 @@
 ## `master`
 - update common dependency to 0.2.3
 - provide more feedback when sending using testhub
+- retry send upon server-busy
+- use a new connection for each sender and receiver
 
 ## `v0.1.0`
 - initial release